requirements.txt
cloudinary==1.41.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
streamlit==1.41.0
toml==0.10.2
